docs(spec): hashes in binary conditions must be length-prefixed
The bitmasks of conditions may be ambiguous. If they contain multiple meta-types (types of
conditions that may contain other conditions), then it can't be determined which is the top-level
type. This is generally fine, since we don't need to know the type. But it does mean that in order
to parse the hash, we need to know its length.

BREAKING CHANGE: Hashes are length-prefixed.