feat: treat types and bitmasks separately
We noticed that assigning a new bit for each type could be very wasteful in the future. For instance
if we added SHA-512, we would likely want to add preimage, prefix and threshold types for it. It
makes sense to group functionality into suites so that each bit can refer to more than one new type.

It also makes sense to overlay suites. For instance, one implementation might not want to allow
threshold types. In order to make that possible, we can make THRESHOLD a feature suite and place
the THRESHOLD-SHA-256 condition type into both the THRESHOLD and SHA-256 feature suites.