feat: remove version field
Conditions and fulfillments both start now with two integer fields that are both evaluated as "if this field has a value higher than I understand, abort". This is redundant. The type field *is* our version field and when we want to change either the binary or string formats, we can simply use a type that has not been assigned.