fix(rsa): fix message padding for moduli with certain bit lengths
OpenSSL expects and returns messages in buffers of the same length as the modulus. However, PSS
returns and expects buffers of the same length as the encoded message. Since the encoded message is
one bit shorter, this breaks if the modulus is exactly one bit larger than a byte boundary, e.g.
1025-bit modulus.

This commit also contains a refactor, moving the RSA signing and verification code out of the RSA condition and into a reusable and testable class.