Better test suite, fix for non-V8 browsers (#53)
* refactor(ed25519): remove unused method

* test: generate data-driven test suite

Setup a harness for generating a data-driven test suite out of this implementation's tests.

* fix: support platforms without Error.captureStackTrace

Some environments, including Phantom, do not have Error.captureStackTrace available. In those cases,
we'll simply skip collecting a stack trace for now.

* test(preimage): add unit tests

* refactor(rsa): remove forceSalt feature and use sinon instead

It's not very nice to have testing-related code in the main library. We remove the forceSalt testing
feature from the library and instead stub crypto.randomBytes using Sinon.

* fix(rsa): fix message padding for moduli with certain bit lengths

OpenSSL expects and returns messages in buffers of the same length as the modulus. However, PSS
returns and expects buffers of the same length as the encoded message. Since the encoded message is
one bit shorter, this breaks if the modulus is exactly one bit larger than a byte boundary, e.g.
1025-bit modulus.

This commit also contains a refactor, moving the RSA signing and verification code out of the RSA condition and into a reusable and testable class.

* test(rsa): add RSA-provided sha1 test vectors and sha256 versions

* refactor: allow subconditions to be passed as objects or strings

* test: add tests for shorthand functions

* test(util): add more tests

* test(condition): add more tests for conditions

* test(ed25519): more tests

* test(condition): add a test for the type registry

This exercises the type registry's sanity check.

* chore: bump oer-tools version

* chore: update wallaby.js

* test(prefix): add tests for setSubcondition